+++ dynmix/dlm.py
@@ -1,0 +1,66 @@
+'''
+Implements simple Dynamic Linear Model routines and also includes
+routines for "univariate" DLMs with varying repeated observations
+at each time instant.
+
+Copyright (c) Victhor S. Sartório. All rights reserved.
+This Source Code Form is subject to the terms of the Mozilla Public
+License, v. 2.0. If a copy of the MPL was not distributed with this
+file, You can obtain one at http://mozilla.org/MPL/2.0/.
+'''
+
+import numpy as np
+
+
+def dlm_filter(Y, F, G, V, W, m0 = None, C0 = None):
+    '''
+    Peforms the basic Kalman filter for a Dynamic Linear Model.
+    '''
+
+    T = Y.shape[0]
+    n, p = F.shape
+
+    if Y.shape[1] != n or V.shape[0] != n or V.shape[1] != n:
+        raise ValueError("Observational dimension mismatch")
+    if G.shape[0] != p or G.shape[1] != p:
+        raise ValueError("G dimension mismatch")
+    if W.shape[0] != p or W.shape[1] != p:
+        raise ValueError("G dimension mismatch")
+
+    if m0 is None:
+        m0 = np.zeros(p)
+    if C0 is None:
+        C0 = np.diag(np.ones(p))
+    
+    a = np.empty((T, p))
+    R = np.empty((T, p, p))
+    f = np.empty((T, n))
+    Q = np.empty((T, n, n))
+    m = np.empty((T, p))
+    C = np.empty((T, p, p))
+
+    Gt = G.T
+    Ft = F.T
+
+    a[0] = np.dot(G, m0)
+    R[0] = np.dot(np.dot(G, C0), Gt) + W
+    f[0] = np.dot(F, a[0])
+    Q[0] = np.dot(np.dot(F, R[0]), Ft) + V
+    e = Y[0] - f[0]
+    Qinv = np.linalg.inv(Q[0])
+    A = np.dot(np.dot(R[0], Ft), Qinv)
+    m[0] = a[0] + np.dot(A, e)
+    C[0] = R[0] - np.dot(np.dot(A, Q[0]), A.T)
+
+    for t in range(1, T):
+        a[t] = np.dot(G, m[t-1])
+        R[t] = np.dot(np.dot(G, C[t-1]), Gt) + W
+        f[t] = np.dot(F, a[t])
+        Q[t] = np.dot(np.dot(F, R[t]), Ft) + V
+        e = Y[t] - f[t]
+        Qinv = np.linalg.inv(Q[t])
+        A = np.dot(np.dot(R[t], Ft), Qinv)
+        m[t] = a[t] + np.dot(A, e)
+        C[t] = R[t] - np.dot(np.dot(A, Q[t]), A.T)
+
+    return m, C, a, R, f, Q